Splitting scream_utils by concept, and distinguishing between scream and tpl's
