+++ components/scream/CMakeLists.txt
@@ -43,9 +43,12 @@ SET (CMAKE_CXX_FLAGS ${KOKKOS_CXXFLAGS_STR})
 include(SetCompilerFlags)
 
 set (SCREAM_LINK_FLAGS ${KOKKOS_LDFLAGS_STR})
-set (SCREAM_INCLUDE_DIRS ${Kokkos_INCLUDE_DIR} ${SCREAM_SRC_DIR})
-set (SCREAM_LIBRARY_DIRS ${Kokkos_LIBRARY_DIR})
-set (SCREAM_LIBRARIES ${KOKKOS_LIBS})
+set (SCREAM_INCLUDE_DIRS ${SCREAM_SRC_DIR})
+set (SCREAM_LIBRARY_DIRS ${CMAKE_CURRENT_BINARY_DIR}/src)
+
+set (SCREAM_TPL_INCLUDE_DIRS ${Kokkos_INCLUDE_DIR})
+set (SCREAM_TPL_LIBRARY_DIRS ${Kokkos_LIBRARY_DIR})
+set (SCREAM_TPL_LIBRARIES ${KOKKOS_LIBS})
 
 
 if (${SCREAM_DOUBLE_PRECISION})
