+++ components/scream/src/physics/p3/p3_f90.cpp
@@ -1,6 +1,10 @@
 #include "p3_f90.hpp"
 #include "p3_ic_cases.hpp"
-#include "share/util/scream_util.hpp"
+
+#include "share/error_defs.hpp"
+#include "share/util/scream_utils.hpp"
+#include "share/util/math_utils.hpp"
+
 
 using scream::Real;
 using scream::Int;
