+++ components/scream/src/physics/p3/p3_ic_cases.cpp
@@ -1,5 +1,6 @@
 #include "p3_ic_cases.hpp"
-#include "share/util/scream_util.hpp"
+#include "share/util/scream_utils.hpp"
+#include "share/error_defs.hpp"
 
 namespace scream {
 namespace p3 {
