+++ components/scream/src/share/scream_macros.hpp
@@ -1,0 +1,21 @@
+#ifndef SCREAM_MACROS_HPP
+#define SCREAM_MACROS_HPP
+
+#if defined __INTEL_COMPILER
+# define vector_ivdep _Pragma("ivdep")
+# ifdef _OPENMP
+#  define vector_simd _Pragma("omp simd")
+# else
+#  define vector_simd _Pragma("simd")
+# endif
+#elif defined __GNUG__
+# define vector_ivdep _Pragma("GCC ivdep")
+# define vector_simd _Pragma("GCC ivdep")
+# define restrict __restrict__
+#else
+# define vector_ivdep
+# define vector_simd
+# define restrict
+#endif
+
+#endif // SCREAM_MACROS_HPP
