+++ components/scream/src/share/tests/CMakeLists.txt
@@ -1,10 +1,10 @@
-link_directories(${SCREAM_LIBRARY_DIRS})
+link_directories(${SCREAM_LIBRARY_DIRS} ${SCREAM_TPL_LIBRARY_DIRS})
 add_executable(util_unittest_f90 unit_test.f90)
 add_executable(util_tests
   util_tests.cpp
   ../util/scream_catch_main.cpp)
 foreach (exe util_tests util_unittest_f90)
-  target_link_libraries(${exe} scream_share ${SCREAM_LIBRARIES})
+  target_link_libraries(${exe} scream_share ${SCREAM_TPL_LIBRARIES})
   target_include_directories(${exe} PUBLIC ${SCREAM_INCLUDE_DIRS} ${CATCH_INCLUDE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/..)
   set_target_properties(${exe} PROPERTIES
     LINK_FLAGS "${SCREAM_LINK_FLAGS}"
