+++ components/scream/src/share/tests/util_tests.cpp
@@ -1,6 +1,6 @@
 #include "catch2/catch.hpp"
 
-#include "share/util/scream_util.hpp"
+#include "share/util/scream_utils.hpp"
 
 namespace {
